VKT(Frontend): OPHVKTKEH-179 some mobile grid refactoring
