+++ frontend/packages/vkt/src/components/publicEnrollment/PublicEnrollmentGrid.tsx
@@ -1,23 +1,18 @@
 import { Grid, Paper } from '@mui/material';
-import { useCallback, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useNavigate, useParams } from 'react-router';
-import {
-  H2,
-  LoadingProgressIndicator,
-  StackableMobileAppBar,
-  Text,
-} from 'shared/components';
+import { LoadingProgressIndicator } from 'shared/components';
 import { APIResponseStatus } from 'shared/enums';
 import { useWindowProperties } from 'shared/hooks';
-import { MobileAppBarState } from 'shared/interfaces';
 
 import { PublicEnrollmentControlButtons } from 'components/publicEnrollment/PublicEnrollmentControlButtons';
 import { PublicEnrollmentExamEventDetails } from 'components/publicEnrollment/PublicEnrollmentExamEventDetails';
 import { PublicEnrollmentPaymentSum } from 'components/publicEnrollment/PublicEnrollmentPaymentSum';
+import { PublicEnrollmentPhoneGrid } from 'components/publicEnrollment/PublicEnrollmentPhoneGrid';
 import { PublicEnrollmentStepContents } from 'components/publicEnrollment/PublicEnrollmentStepContents';
 import { PublicEnrollmentStepper } from 'components/publicEnrollment/PublicEnrollmentStepper';
 import { PublicEnrollmentTimer } from 'components/publicEnrollment/PublicEnrollmentTimer';
-import { useCommonTranslation, usePublicTranslation } from 'configs/i18n';
+import { useCommonTranslation } from 'configs/i18n';
 import { useAppDispatch, useAppSelector } from 'configs/redux';
 import { AppRoutes, EnrollmentStatus } from 'enums/app';
 import { PublicEnrollmentFormStep } from 'enums/publicEnrollment';
@@ -29,7 +24,6 @@ import {
 } from 'redux/reducers/publicEnrollment';
 import { publicEnrollmentSelector } from 'redux/selectors/publicEnrollment';
 import { ExamEventUtils } from 'utils/examEvent';
-import { PublicEnrollmentUtils } from 'utils/publicEnrollment';
 
 export const PublicEnrollmentGrid = ({
   activeStep,
@@ -39,7 +33,6 @@ export const PublicEnrollmentGrid = ({
   // State
   const [isStepValid, setIsStepValid] = useState(false);
   const [showValidation, setShowValidation] = useState(false);
-  const [appBarState, setAppBarState] = useState<MobileAppBarState>({});
 
   // Redux
   const dispatch = useAppDispatch();
@@ -55,9 +48,6 @@ export const PublicEnrollmentGrid = ({
 
   // i18n
   const translateCommon = useCommonTranslation();
-  const { t } = usePublicTranslation({
-    keyPrefix: 'vkt.component.publicEnrollment.stepper',
-  });
 
   const navigate = useNavigate();
   const params = useParams();
@@ -103,15 +93,6 @@ export const PublicEnrollmentGrid = ({
     AppRoutes.PublicEnrollment
   );
 
-  const memoizedSetAppBarState = useCallback(
-    (order: number, height: number) =>
-      setAppBarState((prev) => ({
-        ...prev,
-        [order]: height,
-      })),
-    []
-  );
-
   const isLoading = [
     renewReservationStatus,
     cancelStatus,
@@ -129,115 +110,6 @@ export const PublicEnrollmentGrid = ({
   const isPaymentSumAvailable =
     isPreviewStepActive &&
     (reservationDetails?.reservation || isShiftedFromQueue);
-
-  const getNextEnrollmentStep = () => {
-    const steps = PublicEnrollmentUtils.getEnrollmentSteps(hasReservation);
-    const currentIndex = steps.findIndex((step) => step === activeStep);
-
-    return steps[currentIndex + 1];
-  };
-
-  const renderPhoneView = () => (
-    <>
-      <Grid className="public-enrollment__grid" item>
-        {!isDoneStepActive && (
-          <StackableMobileAppBar
-            order={1}
-            state={appBarState}
-            setState={memoizedSetAppBarState}
-          >
-            <div className="rows">
-              {reservationDetails?.reservation && !isDoneStepActive && (
-                <PublicEnrollmentTimer
-                  reservation={reservationDetails.reservation}
-                  isLoading={isLoading}
-                />
-              )}
-              {isPaymentSumAvailable && (
-                <PublicEnrollmentPaymentSum enrollment={enrollment} />
-              )}
-            </div>
-          </StackableMobileAppBar>
-        )}
-        <Paper elevation={3}>
-          <LoadingProgressIndicator isLoading={isLoading} displayBlock={true}>
-            {selectedExamEvent && (
-              <div
-                className={
-                  isLoading
-                    ? 'dimmed public-enrollment__grid__form-container'
-                    : 'public-enrollment__grid__form-container'
-                }
-              >
-                {!isShiftedFromQueue && (
-                  <div className="columns gapped">
-                    <PublicEnrollmentStepper
-                      activeStep={activeStep}
-                      includePaymentStep={ExamEventUtils.hasOpenings(
-                        selectedExamEvent
-                      )}
-                    />
-                    <div className="rows gapped-xs align-items-center grow">
-                      <H2>
-                        {t(`step.${PublicEnrollmentFormStep[activeStep]}`)}
-                      </H2>
-                      <div>
-                        <Text>
-                          {translateCommon('next')}
-                          {': '}
-                          {t(
-                            `step.${
-                              PublicEnrollmentFormStep[getNextEnrollmentStep()]
-                            }`
-                          )}
-                        </Text>
-                      </div>
-                    </div>
-                  </div>
-                )}
-                <PublicEnrollmentExamEventDetails
-                  examEvent={selectedExamEvent}
-                  showOpenings={hasReservation && !isDoneStepActive}
-                />
-                <PublicEnrollmentStepContents
-                  selectedExamEvent={selectedExamEvent}
-                  activeStep={activeStep}
-                  enrollment={enrollment}
-                  isLoading={isLoading}
-                  setIsStepValid={setIsStepValid}
-                  showValidation={showValidation}
-                />
-              </div>
-            )}
-          </LoadingProgressIndicator>
-        </Paper>
-        {!isDoneStepActive &&
-          !isAuthenticateStepActive &&
-          reservationDetails && (
-            <StackableMobileAppBar
-              order={2}
-              state={appBarState}
-              setState={memoizedSetAppBarState}
-            >
-              <div className="rows">
-                <PublicEnrollmentControlButtons
-                  submitStatus={enrollmentSubmitStatus}
-                  activeStep={activeStep}
-                  enrollment={enrollment}
-                  reservationDetails={reservationDetails}
-                  isLoading={isLoading}
-                  isStepValid={isStepValid}
-                  setShowValidation={setShowValidation}
-                  isPaymentLinkPreviewView={
-                    isShiftedFromQueue && isPreviewStepActive
-                  }
-                />
-              </div>
-            </StackableMobileAppBar>
-          )}
-      </Grid>
-    </>
-  );
 
   const renderDesktopView = () => (
     <>
@@ -316,7 +188,23 @@ export const PublicEnrollmentGrid = ({
       direction="column"
       className="public-enrollment"
     >
-      {isPhone ? renderPhoneView() : renderDesktopView()}
+      {isPhone ? (
+        <PublicEnrollmentPhoneGrid
+          isStepValid={isStepValid}
+          isLoading={isLoading}
+          isShiftedFromQueue={isShiftedFromQueue}
+          isPaymentSumAvailable={!!isPaymentSumAvailable}
+          isAuthenticateStepActive={isAuthenticateStepActive}
+          isPreviewStepActive={isPreviewStepActive}
+          isDoneStepActive={isDoneStepActive}
+          setShowValidation={setShowValidation}
+          setIsStepValid={setIsStepValid}
+          showValidation={showValidation}
+          activeStep={activeStep}
+        />
+      ) : (
+        renderDesktopView()
+      )}
     </Grid>
   );
 };
